Towards a simple menu solution; Need to fix tests
