+++ grr/server/grr_response_server/gui/ui/components/expandable_hash/expandable_hash.scss
@@ -1,3 +1,8 @@
-.bottom-border {
+.show-button {
   border-bottom: 1px dotted rgb(126, 126, 126);
+  cursor: pointer;
+}
+
+.menu-item {
+  padding: 0.3em;
 }
