+++ grr/server/grr_response_server/gui/ui/components/expandable_hash/expandable_hash.ts
@@ -1,7 +1,23 @@
-import { Component, Input, ChangeDetectorRef } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { Hash } from '../../lib/api/api_interfaces'
 
-const MAX_TIME_MOUSE_IN_FLIGHT_MS = 200;
+/** Functionality to aggregate a number of (hashType, hashValue) pairs into a single string */
+export class HashTextAggregator {
+  private textBuffer: string[] = [];
+
+  private append(line: string) {
+    this.textBuffer.push(line);
+  }
+
+  appendHashTypeAndValue(hashType: string, hashValue: string) {
+    const line = `${hashType}: ${hashValue}`;
+    this.append(line);
+  }
+
+  toString(): string {
+    return this.textBuffer.join('\n');
+  }
+}
 
 /**
  * Displays a default text. When the text is hovered, a pop-up appears
@@ -15,57 +31,33 @@ const MAX_TIME_MOUSE_IN_FLIGHT_MS = 200;
 export class ExpandableHash {
   @Input() hashes?: Hash;
 
-  shouldShowPopup = false;
-
-  private mouseInsidePopup = false;
-  private mouseInsideTextArea = false;
-
-  private lastTimeout?: number;
-
-  get atLeastOneHashAvailable() {
-    return (this.hashes?.sha256 ?? this.hashes?.sha1 ?? this.hashes?.md5) !== undefined;
-  }
-
-  constructor(private changeDetectorRef: ChangeDetectorRef) { }
-
-  mouseEnteredTextArea() {
-    this.showPopupNow();
-    this.mouseInsideTextArea = true;
-  }
-  mouseLeftTextArea() {
-    this.schedulePopupHiding();
-    this.mouseInsideTextArea = false;
-  }
-
-  mouseEnteredPopup() {
-    this.showPopupNow();
-    this.mouseInsidePopup = true;
-  }
-  mouseLeftPopup() {
-    this.hidePopupNow(false);
-    this.mouseInsidePopup = false;
-  }
-
-  private showPopupNow() {
-    this.shouldShowPopup = true;
-  }
-
-  private schedulePopupHiding() {
-    if (this.lastTimeout) {
-      window.clearTimeout(this.lastTimeout);
+  get hashesAvailable(): number {
+    if (this.hashes === undefined) {
+      return 0;
     }
 
-    this.lastTimeout = window.setTimeout(() => {
-      if (!this.mouseInsidePopup && !this.mouseInsideTextArea) {
-        this.hidePopupNow(true);
-      }
-    }, MAX_TIME_MOUSE_IN_FLIGHT_MS);
+    return [this.hashes.sha256, this.hashes.sha1, this.hashes.md5]
+      .map(hash => hash ? 1 : 0 as number)
+      .reduce((total, current) => total+current);
   }
 
-  private hidePopupNow(forceChangeDetection: boolean) {
-    this.shouldShowPopup = false;
-    if (forceChangeDetection) {
-      this.changeDetectorRef.detectChanges();
+  get completeHashInformation(): string {
+    const hashText = new HashTextAggregator();
+
+    if (this.hashes === undefined) {
+      return '';
     }
+
+    if (this.hashes.sha256) {
+      hashText.appendHashTypeAndValue('SHA-256', this.hashes.sha256);
+    }
+    if (this.hashes.sha1) {
+      hashText.appendHashTypeAndValue('SHA-1', this.hashes.sha1);
+    }
+    if (this.hashes.md5) {
+      hashText.appendHashTypeAndValue('MD5', this.hashes.md5);
+    }
+
+    return hashText.toString();
   }
 }
