--- grr/server/grr_response_server/gui/ui/components/expandable_hash/hashes/hashes.ng.html
@@ -1,52 +1,0 @@
-<div class="container mat-body-1">
-  <div *ngIf="hashesAvailable === 0" class="no-hashes">
-    No available hashes to display.
-  </div>
-
-  <div *ngIf="hashes?.sha256" class="hash-holder">
-    <button
-      class="copy-button"
-      mat-icon-button
-      aria-label="Copy SHA-256 to clipboard"
-      [cdkCopyToClipboard]="hashes.sha256">
-      <mat-icon class="menu-icon">content_copy</mat-icon>
-    </button>&nbsp;
-    <span class="hash-name">SHA-256</span>
-    <div class="hash-value">{{hashes.sha256}}</div>
-  </div>
-
-  <div *ngIf="hashes?.sha1" class="hash-holder">
-    <button
-      class="copy-button"
-      mat-icon-button
-      aria-label="Copy SHA-1 to clipboard"
-      [cdkCopyToClipboard]="hashes.sha1">
-      <mat-icon class="menu-icon">content_copy</mat-icon>
-    </button>&nbsp;
-    <span class="hash-name">SHA-1</span>
-    <div class="hash-value">{{hashes.sha1}}</div>
-  </div>
-
-  <div *ngIf="hashes?.md5" class="hash-holder">
-    <button
-      class="copy-button"
-      mat-icon-button
-      aria-label="Copy MD5 to clipboard"
-      [cdkCopyToClipboard]="hashes.md5">
-      <mat-icon class="menu-icon">content_copy</mat-icon>
-    </button>&nbsp;
-    <span class="hash-name">MD5</span>
-    <div class="hash-value">{{hashes.md5}}</div>
-  </div>
-
-  <div *ngIf="hashesAvailable > 1" class="copy-all-section">
-    <button
-    class="copy-button"
-      mat-icon-button
-      aria-label="Copy complete hash information to clipboard"
-      [cdkCopyToClipboard]="completeHashInformation">
-      <mat-icon class="menu-icon">content_copy</mat-icon>
-    </button>&nbsp;
-    <span class="copy-all-text">All hash information</span>
-  </div>
-</div>
