--- grr/server/grr_response_server/gui/ui/components/expandable_hash/hashes/hashes.scss
@@ -1,45 +1,0 @@
-@use '~material-theme' as c;
-@import '~@angular/material/theming';
-
-.container {
-  background-color: c.mat-color(c.$background, app-bar);
-
-  border: 1px solid c.mat-color(c.$foreground, divider-light);
-  border-radius: 8px;
-  @include mat-elevation(2);
-
-  font-family: c.$google-sans-family;
-  color: c.mat-color(c.$foreground, text-light);
-
-  padding: 0.5em;
-}
-
-.copy-button {
-  font-size: 16px;
-
-  line-height: 18px;
-  height: 18px;
-  width: 18px;
-
-  mat-icon {
-    font-size: 16px;
-  }
-}
-
-.hash-holder .copy-all-section {
-  margin-top: 0.3em;
-  white-space: nowrap;
-}
-
-.hash-name {
-  font-weight: bold;
-  margin-right: 0.5em;
-}
-
-.hash-value {
-  display: inline-block;
-}
-
-.no-hashes {
-  font-style: italic;
-}
