+++ grr/server/grr_response_server/gui/ui/components/expandable_hash/module.ts
@@ -1,24 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ExpandableHash } from './expandable_hash'
-import { Hashes } from './hashes/hashes'
-import { OverlayModule } from '@angular/cdk/overlay';
 import { ClipboardModule } from '@angular/cdk/clipboard';
-import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
+import { MatMenuModule } from '@angular/material/menu';
 
 
 @NgModule({
   imports: [
     CommonModule,
-    OverlayModule,
     ClipboardModule,
-    MatIconModule,
     MatButtonModule,
+    MatMenuModule,
   ],
   declarations: [
     ExpandableHash,
-    Hashes,
   ],
   exports: [
     ExpandableHash,
